core pipeline v2/requirements.txt
aixplain>=0.2.33
requests>=2.1.0
tqdm>=4.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
validators>=0.20.0
filetype>=1.2.0
click>=7.1.2
PyYAML>=6.0.1
dataclasses-json>=0.5.2
Jinja2==3.1.6
sentry-sdk>=1.0.0
pydantic>=2.10.6
filelock>=3.0.0
PyPDF2>=3.0.1
pypdf>=3.0.0
python-docx>=0.8.11
msal>=1.20.0
nest-asyncio>=1.5.6
